all function is realised but there are still some problems
Problems
1. while focus on RDR2 capturer get some problem.
Only one same frame will be captured each period.
2. SetForegroundWindow only can succeed once.
When focus on console and then focus on RDR2 will can't focus on console
again.
3. Syntheticed video  with mp4 can be watched in windows potplayer.
But can't be watched by windows media and phone.
But video with avi has not those problem.However,avi is very fuzzy.
So I convert images to video with mp4 and then convert mp4 to avi.
It's useful.
But there are also some problem.QQZone and wechat can't send avi video.
The solution is modify .avi to .mp4 in explorer.
Then send it to others in wechat to save the video sent in wechat.
Then we can send it in QQZone.
4. RDR2 will into video camera while we do nothing for 2 minutes.
So I need a backgroudservice to focus on RDR2 and move mouse tiny pixels
Then focus on console again.

TODO
1. we can capture every 0.5s to make video more fluent.
2. Commandline support.
3. suspend capturer
4. Host
5. I should find out why mp4 exists such problems and why avi is fuzzy.
6. write custom images to video method do not use JoinImageSequence.
Because JoinImageSequence  also just use FromFileInput and OutputFile.
So I can write custom method refer to JoinImageSequence.
and so on